+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## [1.x] (2019-06-xx)
+
+[1.x]: https://github.com/sounisi5011/metalsmith-pug-extra/compare/v1.0.1...HEAD
+
+### Features
+
+* pattern options of render() function Issue [#24] / PR [#31] ([9469c55])
+
+[#24]: https://github.com/sounisi5011/metalsmith-pug-extra/issues/24
+[#31]: https://github.com/sounisi5011/metalsmith-pug-extra/pull/31
+[9469c55]: https://github.com/sounisi5011/metalsmith-pug-extra/commit/9469c5525c002015a3f17a8aa3d09cb828599616
+
 ## [1.0.1] (2019-06-16)
 
 [1.0.1]: https://github.com/sounisi5011/metalsmith-pug-extra/compare/v1.0.0...v1.0.1